Rename: flags -> doccompilerflag
