+++ generatedocs/cmd/compiledocs/main.go
@@ -2,11 +2,11 @@ package main
 
 import (
 	"github.com/qlik-oss/gopherciser/generatedocs/pkg/doccompiler"
-	"github.com/qlik-oss/gopherciser/generatedocs/pkg/flags"
+	"github.com/qlik-oss/gopherciser/generatedocs/pkg/doccompilerflag"
 )
 
 func main() {
 	compiler := doccompiler.New()
-	compiler.AddDataFromDir(flags.DataRoot())
-	compiler.CompileToFile(flags.OutputFile())
+	compiler.AddDataFromDir(doccompilerflag.DataRoot())
+	compiler.CompileToFile(doccompilerflag.OutputFile())
 }
